+++ cmd/frontend/internal/db/gen.go
@@ -1,4 +1,4 @@
 package db
 
 // $PGHOST, $PGUSER, $PGPORT etc. must be set to run this generate script.
-//go:generate go run schemadoc/main.go schema.md
+//go:generate sh -c "DISABLE_CONFIG=t go run schemadoc/main.go schema.md"